backend/requirements.txt
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiohttp-retry==2.9.1
aiosignal==1.3.1
aiosmtplib==4.0.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
caio==0.9.24
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
cryptography==45.0.3
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.26.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
minio==7.2.15
miniopy-async==1.23.0
multidict==6.1.0
pgpt_python==0.1.2
propcache==0.3.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
PyJWT==2.10.1
PyMuPDF==1.25.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
returns==0.25.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.38.5
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.20.0
